Remove old code and hardcoded dimensions
