--- app/src/main/res/layout/list_row_pickuser_searchuser.xml
@@ -1,105 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-
-    Wire
-    Copyright (C) 2018 Wire Swiss GmbH
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
--->
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:app="http://schemas.android.com/apk/res-auto"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:orientation="horizontal"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:gravity="center_vertical"
->
-
-    <com.waz.zclient.common.views.ChatheadView
-        android:id="@+id/cv_pickuser__searchuser_chathead"
-        android:layout_width="@dimen/search__chathead__size"
-        android:layout_height="@dimen/search__chathead__size"
-        android:layout_gravity="center"
-        android:layout_marginStart="16dp"
-        android:layout_marginEnd="16dp"
-        app:glyph_size="@dimen/people_picker__chathead__glpyh"
-        app:isSelectable="true"
-        app:show_border="false"
-    />
-
-    <LinearLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:orientation="vertical"
-    >
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="63.5dp"
-            android:gravity="center_vertical"
-            android:layout_marginEnd="16dp"
-            android:orientation="horizontal">
-
-            <com.waz.zclient.usersearch.views.ContactListItemTextView
-                android:id="@+id/clitv__contactlist__user__text_view"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:gravity="start|center_vertical"/>
-
-            <View
-                android:layout_width="0dp"
-                android:layout_height="match_parent"
-                android:layout_weight="1"
-            />
-
-            <com.waz.zclient.ui.text.TypefaceTextView
-                android:id="@+id/guest_indicator"
-                android:visibility="gone"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="@dimen/list_indicator_between_padding"
-                android:paddingStart="@dimen/wire__padding__small"
-                android:paddingEnd="@dimen/wire__padding__small"
-                android:text="@string/people_picker__guest_label"
-                android:textColor="@color/text__primary_light"
-                android:textSize="@dimen/wire__text_size__small"
-                android:gravity="center|end"
-                android:background="@drawable/rounded_corner_background_bubble"
-                app:w_font="@string/wire__typeface__light"
-            />
-
-            <android.support.v7.widget.AppCompatCheckBox
-                android:id="@+id/pick_user_checkbox"
-                android:visibility="gone"
-                android:layout_width="@dimen/share_extension__row__checkbox_size"
-                android:layout_height="@dimen/share_extension__row__checkbox_size"
-                android:layout_gravity="center_vertical"
-                android:button="@drawable/checkbox_black"
-                android:layout_marginStart="@dimen/wire__padding__regular"
-                tools:targetApi="lollipop"/>
-
-        </LinearLayout>
-
-        <View
-            android:id="@+id/separator"
-            android:layout_width="match_parent"
-            android:layout_height="0.5dp"
-            android:background="@color/white_8"
-        />
-
-    </LinearLayout>
-
-</LinearLayout>
