+++ app/src/main/res/layout/single_user_row_view.xml
@@ -22,7 +22,7 @@
               xmlns:app="http://schemas.android.com/apk/res-auto"
               xmlns:tools="http://schemas.android.com/tools"
               android:layout_width="match_parent"
-              android:layout_height="64dp"
+              android:layout_height="@dimen/user_row_height"
     >
 
     <com.waz.zclient.common.views.ChatheadView
@@ -61,8 +61,8 @@
 
         <ImageView
             android:id="@+id/guest_indicator"
-            android:layout_width="16dp"
-            android:layout_height="16dp"
+            android:layout_width="@dimen/user_row_icon_size"
+            android:layout_height="@dimen/user_row_icon_size"
             android:layout_gravity="center_vertical"
             android:layout_marginStart="@dimen/wire__padding__small"
             android:layout_marginEnd="@dimen/wire__padding__small"
@@ -82,8 +82,8 @@
 
         <ImageView
             android:id="@+id/next_indicator"
-            android:layout_width="16dp"
-            android:layout_height="16dp"
+            android:layout_width="@dimen/user_row_icon_size"
+            android:layout_height="@dimen/user_row_icon_size"
             android:layout_gravity="center_vertical"
             android:layout_marginStart="@dimen/wire__padding__small"
             android:layout_marginEnd="@dimen/wire__padding__small"
