--- app/src/main/res/layout/startui_user.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-
-    Wire
-    Copyright (C) 2018 Wire Swiss GmbH
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
--->
-<com.waz.zclient.usersearch.views.SearchResultUserRowView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/srurv_startui_user"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    />
