+++ wire-core/src/main/res/values/dimens.xml
@@ -933,4 +933,9 @@
     <dimen name="invitation_list_item_height">48dp</dimen>
     <dimen name="add_remove_service_button_height">40dp</dimen>
     <dimen name="integration_details_picture_size">64dp</dimen>
+    <dimen name="create_conv_input_margin_top">40dp</dimen>
+    <dimen name="menu_item_height">64dp</dimen>
+    <dimen name="guest_options_icon_size">16dp</dimen>
+    <dimen name="user_row_height">64dp</dimen>
+    <dimen name="user_row_icon_size">16dp</dimen>
 </resources>
