+++ wire-core/src/main/res/values/strings.xml
@@ -1315,5 +1315,6 @@
 
     <string name="open_this_conversation_to_people_outside_your_team">Open this conversation to people outside your team</string>
     <string name="allow_guests">Allow Guests</string>
+    <string name="guest_options_title">Guest Options</string>
 
 </resources>
